backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-jwt-extended==4.7.1
flask-mail==0.10.0
flask-socketio==5.3.6
psycopg2-binary==2.9.10
python-dotenv==1.1.0
twilio==9.4.0
speedtest-cli==2.1.3
chromadb>=0.4.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
statsmodels>=0.14.0
prophet>=1.1.0
